Adding submodules for base components
The RISC-V Core, Boot ROM, and Serial Peripheral should be most of what
we need to start bringing up a minimal system.

The RISC-V Core should be obvious - it's our CPU. Initially, it will be
connected to a simplified memory bus containing our other items.

The Boot ROM will be built from source, loaded into the FPGA's block
RAM, and attached to the memory bus at 0x00000000.

Finally, the Serial Peripheral will provide the main console so we can
see things working and interact with the program provided by the Boot
ROM. This will be a fairly minimal implementation to start with.